PUBDEV-5820: refactored Jetty-dependent classes so that they contain minimum of implementation code - new package water.server.jetty contains those really dependeny on Jetty - new package water.server contains utilities formerly present in JettyHTTPD which actually are not related to Jetty at all
