+++ h2o-core/src/main/java/water/api/PutKeyServlet.java
@@ -2,9 +2,9 @@ package water.api;
 
 import org.apache.commons.io.IOUtils;
 import water.DKV;
-import water.JettyHTTPD;
 import water.Key;
 import water.Value;
+import water.server.ServletUtils;
 
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
@@ -24,7 +24,7 @@ public class PutKeyServlet extends HttpServlet {
 
   @Override
   protected void doPost(HttpServletRequest request, HttpServletResponse response) {
-    String uri = JettyHTTPD.getDecodedUri(request);
+    String uri = ServletUtils.getDecodedUri(request);
     
     try {
       String destKey = paramDestinationKey(request, response);
@@ -33,7 +33,7 @@ public class PutKeyServlet extends HttpServlet {
         return;
       }
 
-      InputStream is = JettyHTTPD.extractPartInputStream(request, response);
+      InputStream is = ServletUtils.extractPartInputStream(request, response);
       if (is == null) {
         return;
       }
@@ -57,9 +57,9 @@ public class PutKeyServlet extends HttpServlet {
       response.setContentType("application/json");
       response.getWriter().write(responsePayload);
     } catch (Exception e) {
-      JettyHTTPD.sendErrorResponse(response, e, uri);
+      ServletUtils.sendErrorResponse(response, e, uri);
     } finally {
-      JettyHTTPD.logRequest("POST", request, response);
+      ServletUtils.logRequest("POST", request, response);
     }
   }
 
@@ -75,7 +75,7 @@ public class PutKeyServlet extends HttpServlet {
 
   private boolean validate(String destKey, Boolean overwrite, HttpServletResponse response) throws IOException {
     if (destKey == null) {
-      JettyHTTPD.sendResponseError(response, HttpServletResponse.SC_BAD_REQUEST, "The field 'destination_frame` is compulsory!");
+      ServletUtils.sendResponseError(response, HttpServletResponse.SC_BAD_REQUEST, "The field 'destination_frame` is compulsory!");
       return false;
     }
     return true;
