+++ h2o-core/src/main/java/water/server/jetty/DelegatingAuthenticator.java
@@ -1,4 +1,4 @@
-package water;
+package water.server.jetty;
 
 import org.eclipse.jetty.security.Authenticator;
 import org.eclipse.jetty.security.ServerAuthException;
