+++ h2o-core/src/test/java/water/api/CustomHttpFilterTest.java
@@ -7,8 +7,8 @@ import org.apache.log4j.spi.LoggingEvent;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import water.H2O;
-import water.JettyHTTPD;
 import water.TestUtil;
+import water.server.jetty.JettyHTTPD;
 
 import javax.servlet.ServletOutputStream;
 import javax.servlet.WriteListener;
